requirements.txt
# Core ML/AI
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.5
openai>=1.0.0
tiktoken>=0.5.0
transformers>=4.36.0
datasets>=2.14.0
accelerate>=0.25.0
bitsandbytes>=0.41.0
peft>=0.7.0

# RAG / Embeddings (using Pinecone - FREE tier available)
pinecone-client>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4        # Alternative: local vector store
chromadb>=0.4.0         # Alternative: local vector store

# Astronomy Libraries
astropy>=6.0.0
astroquery>=0.4.6
photutils>=1.10.0
sep>=1.2.0
reproject>=0.11.0
regions>=0.7
pyvo>=1.4.0
astroml>=1.0.0
dustmaps>=1.0.0
extinction>=0.4.0

# NASA/ESA Archive Access
astrocut>=0.10
s3fs>=2023.0.0        # AWS S3 for NASA data
requests>=2.28.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.11.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0
aplpy>=2.1.0          # Astronomical plotting

# Image Processing
pillow>=10.0.0
scikit-image>=0.22.0
opencv-python>=4.8.0

# Utilities
pyyaml>=6.0.0
python-dotenv>=1.0.0
tqdm>=4.66.0
rich>=13.0.0

# API Clients (for fine-tuning deployment)
together>=0.2.0       # Together.ai
fireworks-ai>=0.10.0  # Fireworks.ai

# Fine-tuning tools
llama-factory>=0.5.0  # If using LlamaFactory
unsloth>=2024.0       # Fast fine-tuning (optional)

# Development
jupyter>=1.0.0
ipywidgets>=8.1.0
black>=23.0.0
pytest>=7.4.0
sentence-transformers>=2.2.0
